Support IPv6 in address check lookups.
This supports IPv6 in address DNS lookups and socket bind checks. Addresses are
looked up with the system getaddrinfo(), and the v6/v4 preference of this lookup
is respected. IPv6 address literals in the config file are expected to be
bracketed in [], and the result of lookups will be bracketed as well, to support
the ZeroMQ URI format.