Fix more(/all?) remaining instances of #4733 (`not X == Y` -> `X != Y`)
When greping for instances, I had used an overly restrictive regex so that
the list would have few enough false positives (and thus be short enough)
to make manual review feasible.
This time, I used a looser regex to find all(?) remaining instances.
The # of resulting hits was fortunately manageable!