Handle invalid master address in minion config
Gracefully exit the minion while informing the user of the cause of the error if the IP address of the master is malformed such that it can not even be dealt with by Python's socket module.

Unit testing for the above condition.

Closes #9440.

Remove suprious debugging statement.

Remove more spurious debugging.